requirements.txt
fastapi>=0.95
uvicorn[standard]>=0.18
sqlalchemy>=1.4
python-dotenv>=0.21
psycopg2-binary>=2.9
pydantic>=1.10
requests>=2.28
pytest>=7.0
httpx>=0.24
typing-extensions>=4.0
